+++ src/main/kotlin/io/snyk/plugin/ui/toolwindow/SnykTreeCellRenderer.kt
@@ -11,6 +11,7 @@ import io.snyk.plugin.getSnykCachedResults
 import io.snyk.plugin.pluginSettings
 import io.snyk.plugin.snykcode.core.AnalysisData
 import io.snyk.plugin.snykcode.core.PDU
+import io.snyk.plugin.snykcode.core.SnykCodeFile
 import io.snyk.plugin.snykcode.getSeverityAsEnum
 import io.snyk.plugin.ui.PackageManagerIconProvider
 import io.snyk.plugin.ui.getDisabledIcon
@@ -25,6 +26,7 @@ import io.snyk.plugin.ui.toolwindow.nodes.root.RootSecurityIssuesTreeNode
 import io.snyk.plugin.ui.toolwindow.nodes.secondlevel.ErrorTreeNode
 import io.snyk.plugin.ui.toolwindow.nodes.secondlevel.FileTreeNode
 import io.snyk.plugin.ui.toolwindow.nodes.secondlevel.SnykCodeFileTreeNode
+import snyk.common.ProductType
 import snyk.common.SnykError
 import snyk.container.ContainerIssue
 import snyk.container.ContainerIssuesForImage
@@ -66,9 +68,9 @@ class SnykTreeCellRenderer : ColoredTreeCellRenderer() {
                 }
             }
             is FileTreeNode -> {
-                val ossVulnerabilitiesForFile = value.userObject as OssVulnerabilitiesForFile
-                nodeIcon = PackageManagerIconProvider.getIcon(ossVulnerabilitiesForFile.packageManager.toLowerCase())
-                text = ossVulnerabilitiesForFile.sanitizedTargetFile
+                val fileVulns = value.userObject as OssVulnerabilitiesForFile
+                nodeIcon = PackageManagerIconProvider.getIcon(fileVulns.packageManager.toLowerCase())
+                text = fileVulns.sanitizedTargetFile + ProductType.OSS.getCountText(value.childCount)
 
                 val snykCachedResults = getSnykCachedResults(value.project)
                 if (snykCachedResults?.currentOssResults == null) {
@@ -85,14 +87,15 @@ class SnykTreeCellRenderer : ColoredTreeCellRenderer() {
                     if (suggestion.title.isNullOrEmpty()) suggestion.message else suggestion.title
                 }"
                 val parentFileNode = value.parent as SnykCodeFileTreeNode
-                if (!AnalysisData.instance.isFileInCache(parentFileNode.userObject)) {
+                val file = (parentFileNode.userObject as Pair<SnykCodeFile, ProductType>).first
+                if (!AnalysisData.instance.isFileInCache(file)) {
                     attributes = SimpleTextAttributes.GRAYED_ATTRIBUTES
                     nodeIcon = getDisabledIcon(nodeIcon)
                 }
             }
             is SnykCodeFileTreeNode -> {
-                val file = value.userObject
-                text = PDU.toSnykCodeFile(file).virtualFile.name
+                val (file, productType) = value.userObject as Pair<SnykCodeFile, ProductType>
+                text = PDU.toSnykCodeFile(file).virtualFile.name + productType.getCountText(value.childCount)
                 val psiFile = PDU.toPsiFile(file)
                 nodeIcon = psiFile?.getIcon(Iconable.ICON_FLAG_READ_STATUS)
                 if (!AnalysisData.instance.isFileInCache(file)) {
@@ -104,7 +107,7 @@ class SnykTreeCellRenderer : ColoredTreeCellRenderer() {
             is IacFileTreeNode -> {
                 val iacVulnerabilitiesForFile = value.userObject as IacIssuesForFile
                 nodeIcon = PackageManagerIconProvider.getIcon(iacVulnerabilitiesForFile.packageManager.toLowerCase())
-                text = iacVulnerabilitiesForFile.targetFile
+                text = iacVulnerabilitiesForFile.targetFile + ProductType.IAC.getCountText(value.childCount)
 
                 val snykCachedResults = getSnykCachedResults(value.project)
                 if (snykCachedResults?.currentIacResult == null || iacVulnerabilitiesForFile.obsolete) {
@@ -116,7 +119,7 @@ class SnykTreeCellRenderer : ColoredTreeCellRenderer() {
             is ContainerImageTreeNode -> {
                 val issuesForImage = value.userObject as ContainerIssuesForImage
                 nodeIcon = SnykIcons.CONTAINER_IMAGE
-                text = issuesForImage.imageName
+                text = issuesForImage.imageName + ProductType.CONTAINER.getCountText(value.childCount)
 
                 val snykCachedResults = getSnykCachedResults(value.project)
                 if (snykCachedResults?.currentContainerResult == null || issuesForImage.obsolete) {