+++ src/main/kotlin/io/snyk/plugin/ui/toolwindow/nodes/leaf/SuggestionTreeNode.kt
@@ -11,6 +11,7 @@ import io.snyk.plugin.ui.toolwindow.nodes.secondlevel.SnykCodeFileTreeNode
 import io.snyk.plugin.ui.toolwindow.panels.IssueDescriptionPanelBase
 import io.snyk.plugin.ui.toolwindow.panels.SuggestionDescriptionPanel
 import snyk.analytics.IssueInTreeIsClicked
+import snyk.common.ProductType
 import javax.swing.tree.DefaultMutableTreeNode
 
 class SuggestionTreeNode(
@@ -28,8 +29,9 @@ class SuggestionTreeNode(
                 .severity(suggestion.getIssueSeverityOrNull())
                 .build()
         )
-        val snykCodeFile = (this.parent as? SnykCodeFileTreeNode)?.userObject as? SnykCodeFile
+        val snykCodeFileTreeNode = this.parent as? SnykCodeFileTreeNode
             ?: throw IllegalArgumentException(this.toString())
+        val snykCodeFile = (snykCodeFileTreeNode.userObject as Pair<SnykCodeFile, ProductType>).first
         return SuggestionDescriptionPanel(snykCodeFile, suggestion, rangeIndex)
     }
 }